Refactored popup into its own component (for repeated use in leaderboard and profile page)
